+++ src/main/java/subway/view/OutputView.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.view;
 
 import subway.domain.Station;
